+++ rebootpy/client.py
@@ -1805,7 +1805,7 @@ class BasicClient:
 
     async def fetch_ranked_stats(self,
                                  user_id: str,
-                                 season: Seasons
+                                 season: Seasons = None
                                  ) -> List[CompetitiveRank]:
         """|coro|
 
@@ -1837,7 +1837,9 @@ class BasicClient:
         user_id: :class:`str`
             The id of the user you want to fetch stats for.
         season: :class:`Seasons`
-            The season that you want to get ranks from.
+            The season that you want to get ranks from, if not provided it'll get the
+            current seasons ranked tracks automatically.
+            *Defaults to None*
 
         Raises
         ------
@@ -1850,12 +1852,16 @@ class BasicClient:
             A list of all of the users ranks in the requested season.
         """  # noqa
 
+        tracks = season.value if season else tuple(
+            track['trackguid'] for track in await self.http.get_active_tracks()
+        )
+
         raw_tracks = await self.http.get_ranked_tracks(
             user_id=user_id
         )
 
         return [CompetitiveRank(data=track) for track in raw_tracks
-                if track['trackguid'] in season.value]
+                if track['trackguid'] in tracks]
 
     async def fetch_br_stats(self, user_id: str, *,
                              start_time: Optional[DatetimeOrTimestamp] = None,