+++ rebootpy/user.py
@@ -250,7 +250,7 @@ class UserBase:
         )
 
     async def fetch_ranked_stats(self,
-                                 season: Seasons
+                                 season: Seasons = None
                                  ) -> List['CompetitiveRank']:
         """|coro|
 
@@ -280,7 +280,9 @@ class UserBase:
         Parameters
         ----------
         season: :class:`Seasons`
-            The season that you want to get ranks from.
+            The season that you want to get ranks from, if not provided it'll get the
+            current seasons ranked tracks automatically.
+            *Defaults to None*
 
         Raises
         ------